servers/jira-helper/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-helper"
version = "2.1.0"
description = "A Jira and Confluence helper MCP server"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "MCP Server Team", email = "example@example.com"},
]
dependencies = [
    # Core MCP
    "mcp>=1.27.0",
    "mcp-commons>=2.1.1",
    
    # Jira/Confluence integration 
    "atlassian-python-api>=4.0.7",
    
    # HTTP transport
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    
    # Configuration
    "python-dotenv>=1.2.2",
    "PyYAML>=6.0.3",
    "requests>=2.32.5",
    
    # Graph visualization
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0", 
    "pytest-mock>=3.14.0",
]

[project.scripts]
jira-helper = "main:main"

[tool.setuptools]
py-modules = ["main", "config", "tool_config", "jira_client", "exceptions"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
